Problem: zeb_handler cannot deal with multiple accepts values

+++ src/xrap_msg.c
@@ -76,7 +76,7 @@ struct _xrap_msg_t {
 
 //  Put a 1-byte number to the frame
 #define PUT_NUMBER1(host) { \
-    *(byte *) self->needle = (host); \
+    *(byte *) self->needle = (byte) (host); \
     self->needle++; \
 }
